requirements.txt
aiogram>=2.15
pymongo>=3.12.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
logging>=0.4.9
requests>=2.25.1
asyncio>=3.4.3 